chore: update chat imports to be relative, minor code refactoring
