+++ crates/q_cli/src/cli/chat/context.rs
@@ -759,7 +759,7 @@ mod tests {
     use std::io::Stdout;
 
     use super::*;
-    use crate::cli::chat::hooks::HookTrigger;
+    use super::super::hooks::HookTrigger;
 
     // Helper function to create a test ContextManager with Context
     pub async fn create_test_context_manager() -> Result<ContextManager> {
